[#46] refactor: 코드리뷰 반영
- Web3jConfig에 타임아웃, 연결 및 설정값 검증 로직 추가
- Request DTO에 Jakarta Validation 어노테이션 적용
- JPA 엔티티 및 리포지토리 최적화 (타입 변경, nullable 제약, 검증 메서드, 페이지네이션)
- 읽기 전용 서비스 메서드에 @Transactional(readOnly = true) 적용

+++ src/main/java/com/masterpiece/IPiece/blockchain/domain/BlockchainTransaction.java
@@ -40,7 +40,7 @@ public class BlockchainTransaction extends BaseEntity {
     private String tokenAddress;
 
     @Column(name = "block_number", nullable = false)
-    private Integer blockNumber;
+    private Long blockNumber;
 
     @Column(name = "block_hash", length = 66)
     private String blockHash;
